fix: ci error
